refactor: reduce imports of Pages from components
