+++ frontend/src/board/pgn/boardTools/boardButtons/BoardButtons.tsx
@@ -1,6 +1,6 @@
+import { useGame } from '@/games/view/GameContext';
 import { useLightMode } from '@/style/useLightMode';
 import { Box, Paper, Stack } from '@mui/material';
-import { useGame } from '../../../../games/view/GamePage';
 import { useChess } from '../../PgnBoard';
 import ControlButtons from './ControlButtons';
 import StartButtons from './StartButtons';
