+++ frontend/src/board/pgn/boardTools/underboard/comments/CommentEditor.tsx
@@ -1,3 +1,4 @@
+import { useGame } from '@/games/view/GameContext';
 import { Send } from '@mui/icons-material';
 import { CircularProgress, IconButton, Stack, TextField, Tooltip } from '@mui/material';
 import React, { useEffect, useRef, useState } from 'react';
@@ -5,7 +6,6 @@ import { useApi } from '../../../../../api/Api';
 import { RequestSnackbar, useRequest } from '../../../../../api/Request';
 import { useRequiredAuth } from '../../../../../auth/Auth';
 import { PositionComment } from '../../../../../database/game';
-import { useGame } from '../../../../../games/view/GamePage';
 import { BlockBoardKeyboardShortcuts, useChess } from '../../../PgnBoard';
 
 export interface CommentEditorProps {
