+++ frontend/src/board/pgn/boardTools/underboard/comments/Comments.tsx
@@ -1,3 +1,4 @@
+import { useGame } from '@/games/view/GameContext';
 import { Chess, EventType, Move } from '@jackstenglein/chess';
 import {
     Button,
@@ -11,7 +12,6 @@ import {
 import React, { createContext, useContext, useEffect, useState } from 'react';
 import { useLocalStorage } from 'usehooks-ts';
 import { Game, PositionComment } from '../../../../../database/game';
-import { useGame } from '../../../../../games/view/GamePage';
 import { useReconcile } from '../../../../Board';
 import { useChess } from '../../../PgnBoard';
 import Comment from './Comment';
