+++ frontend/src/board/pgn/boardTools/underboard/comments/ReplyEditor.tsx
@@ -1,3 +1,4 @@
+import { useGame } from '@/games/view/GameContext';
 import { LoadingButton } from '@mui/lab';
 import { Button, Stack, TextField } from '@mui/material';
 import { useState } from 'react';
@@ -5,7 +6,6 @@ import { useApi } from '../../../../../api/Api';
 import { RequestSnackbar, useRequest } from '../../../../../api/Request';
 import { useRequiredAuth } from '../../../../../auth/Auth';
 import { PositionComment } from '../../../../../database/game';
-import { useGame } from '../../../../../games/view/GamePage';
 import { BlockBoardKeyboardShortcuts } from '../../../PgnBoard';
 
 interface ReplyEditorProps {
