+++ frontend/src/scoreboard/CohortIcon.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { AllInclusive } from '@mui/icons-material';
 import {
     Tooltip as MuiTooltip,
@@ -8,8 +10,7 @@ import {
 import { SxProps, styled } from '@mui/material/styles';
 import Image from 'next/image';
 import { CSSProperties } from 'react';
-import { ALL_COHORTS } from '../database/user';
-import { MastersCohort } from '../games/list/ListGamesPage';
+import { ALL_COHORTS, MastersCohort } from '../database/user';
 
 export const cohortIcons: Record<string, string> = {
     '0-300': 'https://chess-dojo-images.s3.amazonaws.com/icons/v3/0-300.png',
